## Summary of the changes / Why this improves CrateDB
Fixes:
```
cr> show create table t;
+-----------------------------------------------------+
| SHOW CREATE TABLE doc.t                             |
+-----------------------------------------------------+
| CREATE TABLE IF NOT EXISTS "doc"."t" (              |
|    "a" ARRAY(INTEGER)                               |
| )                                                   |
| CLUSTERED INTO 4 SHARDS                             |

cr> select * from t;
+--------+
| a      |
+--------+
| [null] |
| NULL   |
| [1]    |
| []     |
+--------+

cr> select * from t where a != [];
+-----+
| a   |
+-----+
| [1] | -- missing [null]
+-----+
```


```
cr> show create table t5;
+-----------------------------------------------------+
| SHOW CREATE TABLE doc.t5                            |
+-----------------------------------------------------+
| CREATE TABLE IF NOT EXISTS "doc"."t5" (             |
|    "a" ARRAY(INTEGER) STORAGE WITH (                |
|       columnstore = false                           |
|    )                                                |
| )                                                   |
| CLUSTERED INTO 4 SHARDS                             |

cr> select * from t5;
+--------+
| a      |
+--------+
| []     |
| NULL   |
| [1]    |
| [null] |
+--------+

cr> select * from t5 where a != [1];
+--------+
| a      |
+--------+
| []     |
| NULL   | -- unfiltered
| [null] |
+--------+
```

## Checklist

 - [x] Added an entry in the latest `docs/appendices/release-notes/<x.y.0>.rst` for user facing changes
 - [x] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in the latest `docs/appendices/release-notes/<x.y.0>.rst`
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)