@jeeminso 2024-07-08
The performance impact of the fix:
```
# Results (server side duration in ms)
V1: 5.9.0-d6cc2a6ac021cfcd2a0460e6325e7eda31abb4bf
V2: 5.9.0-39876d7270893b6cc1fe9ee820dd6ee4989ccf0e

// without doc-values

Q: select count(*) from t_arrays_biased_empty where xs != []
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |       33.620 ±  116.033 |     22.729 |     23.859 |     24.952 |   2304.439 |
|   V2    |      180.239 ±  226.668 |    155.052 |    159.129 |    162.733 |   3573.573 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               + 137.12%                           + 147.85%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 137.12%
The test has statistical significance

Q: select count(*) from t_arrays_biased_5_elements where xs != []
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |      389.643 ±  312.787 |    347.844 |    356.771 |    367.616 |   4260.982 |
|   V2    |      586.921 ±  639.280 |    515.331 |    524.726 |    533.100 |   8040.258 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               +  40.40%                           +  38.11%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 40.40%
The test has statistical significance

Q: select count(*) from t_arrays_biased_empty where xs != [1]
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |        1.445 ±    2.217 |      0.878 |      1.241 |      1.382 |     37.635 |
|   V2    |      142.905 ±    4.687 |    138.615 |    141.455 |    143.867 |    182.915 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               + 196.00%                           + 196.52%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 196.00%
The test has statistical significance

Q: select count(*) from t_arrays_biased_5_elements where xs != [1]
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |        1.052 ±    1.297 |      0.573 |      0.761 |      0.830 |     12.115 |
|   V2    |      274.850 ±    9.905 |    264.713 |    270.930 |    278.053 |    347.478 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               + 198.47%                           + 198.88%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 198.47%
The test has statistical significance

// with doc-values

Q: select count(*) from t_arrays_biased_empty_2 where xs != []
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |        3.925 ±   27.227 |      1.575 |      2.516 |      2.639 |    608.619 |
|   V2    |      387.925 ±  182.108 |    357.766 |    366.800 |    373.558 |   2393.549 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               + 195.99%                           + 197.27%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 195.99%
The test has statistical significance

Q: select count(*) from t_arrays_biased_5_elements_2 where xs != []
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |       15.350 ±  163.171 |      5.353 |      8.620 |      8.856 |   3656.509 |
|   V2    |       70.407 ±  248.260 |     50.109 |     51.425 |     52.483 |   4046.039 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               + 128.40%                           + 142.57%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 128.40%
The test has statistical significance

Q: select count(*) from t_arrays_biased_empty_2 where xs != [1]
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |      114.144 ±   73.317 |    101.786 |    105.049 |    108.884 |    987.184 |
|   V2    |      183.272 ±    6.710 |    176.456 |    180.837 |    185.450 |    214.075 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               +  46.49%                           +  53.02%   
There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 46.49%
The test has statistical significance

Q: select count(*) from t_arrays_biased_5_elements_2 where xs != [1]
C: 1
| Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
|   V1    |       48.799 ±  159.281 |     35.002 |     36.880 |     44.718 |   3584.004 |
|   V2    |       26.248 ±    2.330 |     24.629 |     25.381 |     26.433 |     40.712 |
├---------┴-------------------------┴------------┴------------┴------------┴------------┘
|               -  60.10%                           -  36.94%   
There is a 99.84% probability that the observed difference is not random, and the best estimate of that difference is 60.10%
The test has statistical significance
```